Update FRC_Long_version.py
Changing some spelling and grammar bits.

Was going to play with code, but I don't think that'd be a good idea quite yet until I understand it all.
If something crashes after bidding, check line 538 "person_bidding". I changed it from "person_biding" to "person_bidding" but maybe this was a variable and not a string. I don't see any references elsewhere, though.